this adds commonjs like module support to the ender library.
Now all microlibs (which use commonjs exports) can be accessed by a require method.

   var _ = require('underscore'); //returns underscore object

This is particularlly useful when methods step on each other on global $ object.

This also gets backbone working again! hooray!

This also cause many libs to not polute the global scope. This is a breaking change for many bridge files
which rely on globals. These bridges should be update to use the require syntax