Allow `ender build` command to accept a npm json to take modules from
If you provide a path to a json file, `ender build` expects it to be
either a `package.json` or a `npm-shrinkwrap.json`[1] files.
The filename is not important, but the format is, that is to say, if
you provide a `package.json`-like file it must contain a json object
with a `dependencies` hash.

This feature can be useful when trying to lock down dependency versions
between packages. NPM has created a new `shrinkwrap` command to
produce a json file with a list of exact versions to use. Providing
the generated `npm-shrinkwrap.json` file to `ender build` will create
an ender file with just those packages.

[1] http://npmjs.org/doc/shrinkwrap.html